turtlesoup/requirements.txt
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
tqdm>=4.65.0
numpy>=1.24.0
scikit-learn>=1.3.0
rich>=13.7.0
